MNES-1238: MR Suggestions (rename MnestixConnectionServerActions)
